some gruping of CODEOWNERS functionality
