+++ src/commands/generate.ts
@@ -1,59 +1,12 @@
-import { sync } from 'fast-glob';
-import fs from 'fs';
-import { stripIndents } from 'common-tags';
-import { Command, getGlobalOptions } from '../utils/getGlobalOptions';
-import { OUTPUT, INCLUDES, SUCCESS_SYMBOL, SHRUG_SYMBOL, GENERATED_FILE_LEGEND } from '../utils/constants';
 import ora from 'ora';
-import { promisify } from 'util';
-import path, { join, dirname } from 'path';
-import isValidGlob from 'is-valid-glob';
+import path from 'path';
+import { sync } from 'fast-glob';
+import { Command, getGlobalOptions } from '../utils/getGlobalOptions';
+import { OUTPUT, INCLUDES, SUCCESS_SYMBOL, SHRUG_SYMBOL } from '../utils/constants';
+import { ownerRule, createOwnersFile, loadCodeOwnerFiles } from '../utils/codeowners';
 
 type Generate = (options: GenerateInput) => Promise<ownerRule[]>;
 type GenerateInput = { rootDir: string; verifyPaths?: boolean; useMaintainers?: boolean; includes?: string[] };
-
-const readFile = promisify(fs.readFile);
-
-type ownerRule = {
-  filePath: string;
-  owners: string[];
-  glob: string;
-};
-
-const createMatcherCodeownersRule = (filePath: string, rule: string) => {
-  const parts = rule.split(/\s+/);
-  const [glob, ...owners] = parts;
-
-  if (owners.length && isValidGlob(glob)) {
-    return {
-      glob: join(dirname(filePath), glob),
-      owners,
-    };
-  } else {
-    throw new Error(`${rule} in ${filePath} can not be parsed`);
-  }
-};
-const parseCodeOwner = (filePath: string, codeOwnerContent: string): ownerRule[] => {
-  const content = codeOwnerContent.split('\n');
-
-  // TODO: include comments optionally.
-  const filteredRules = content.filter((line) => line && !line.startsWith('#'));
-
-  return filteredRules.map((rule) => ({ filePath, ...createMatcherCodeownersRule(filePath, rule) }));
-};
-
-const loadCodeOwnerFiles = async (dirname: string, files: string[]): Promise<ownerRule[]> => {
-  const codeOwners = await Promise.all(
-    files.map(async (filePath) => {
-      const rawContent = await readFile(filePath);
-
-      const content = rawContent.toString();
-
-      return parseCodeOwner(filePath.replace(`${dirname}/`, ''), content);
-    })
-  );
-
-  return codeOwners.reduce((memo, rules) => [...memo, ...rules], []);
-};
 
 export const generate: Generate = async ({ rootDir, includes = INCLUDES }) => {
   const files = sync(includes, {
@@ -71,16 +24,6 @@ export const generate: Generate = async ({ rootDir, includes = INCLUDES }) => {
   }
 };
 
-const createOwnersFile = (outputFile: string, ownerRules: ownerRule[]): void => {
-  const content = ownerRules.map(
-    (rule) => stripIndents` 
-    # Rule extracted from ${rule.filePath}
-    ${rule.glob} ${rule.owners.join(' ')}
-    `
-  );
-
-  fs.writeFileSync(outputFile, [GENERATED_FILE_LEGEND, ...content].join('\n'));
-};
 interface CommandGenerate extends Command {
   output?: string;
   verifyPaths?: boolean;
@@ -99,12 +42,12 @@ export const command = async (command: CommandGenerate): Promise<void> => {
 
     const outputFile = output || path.join(__dirname, OUTPUT);
 
-    const includes = globalOptions.includes?.length ? globalOptions.includes : INCLUDES;
     if (ownerRules.length) {
       createOwnersFile(outputFile, ownerRules);
 
       loader.stopAndPersist({ text: `CODEOWNERS file was created! location: ${outputFile}`, symbol: SUCCESS_SYMBOL });
     } else {
+      const includes = globalOptions.includes?.length ? globalOptions.includes : INCLUDES;
       loader.stopAndPersist({
         text: `We couldn't find any codeowners under ${includes.join(', ')}`,
         symbol: SHRUG_SYMBOL,
